アプリケーションサービスを共通とLLM用に分割
